feat : LineInfo 삭제 및 Line 객체 변경
