--- src/main/java/subway/domain/LineRepository.java
@@ -1,4 +1,0 @@
-package subway.domain;
-
-public interface LineRepository {
-}
