+++ src/main/java/subway/domain/SectionRepository.java
@@ -1,0 +1,4 @@
+package subway.domain;
+
+public interface SectionRepository {
+}
